Fixed normalization and label calcs
Removed unused commented sections. Added some more, better comments. Changed it so it normalizes at the end for both better math, and to make the date fields normalized. Fixed how the buy, hold, and sell labels were calculated so that they're actually correct now (e.g. use the price of the datapoint instead of the close price for the previous day which was just wrong).