Added exponential generator and noise function
Added an untested exponential generator to DataGen and a windowed noise application function that is tested now in testUtilities. I noticed that the way I'm applying a standard deviation to the data uses the whole average so it ends up applying a huge offset. I probably want to apply noise in a different way.